♻️ 홈베이스 예약 반환 DTO 중첩 클래스 분리
